2019: Allow interrupting the IntcodeEmulator during callbacks
This replaces the need for using `run_until_output` or single-stepping
insructions to be able to get the emulator to give up control.

To do this, a new mutable `context` parameter needs to be added to
all contexts to allow the callback to set the interrupt state.

This could be used to further optimize other days (e.g. Day 23).